requirements.txt
pandas
numpy
scikit-learn==1.6.1
xgboost
lightgbm
catboost
joblib
python-dotenv
google-genai
fastapi==0.115.0
starlette>=0.37.2,<0.39.0
uvicorn
pytest
httpx
